[WIP] Cleaned daemon module. Added support for SIGINT signal.

+++ centralreport/centralreport.py
@@ -33,18 +33,25 @@ class CentralReport(Daemon):
     # Sigterm signal
     SIGTERM_SENT = False
 
-    def sigterm_handler(self, signum, frame):
+    def signal_handler(self, signum, frame):
         """
-            Receives the "SIGTERM" signal from the OS.
-            In this case, we must CentralReport immediately.
+            Receives signals from the OS.
+
         """
-        if not self.SIGTERM_SENT:
-            self.SIGTERM_SENT = True  # Prevents if SIGTERM is received twice.
+        if signum == signal.SIGTERM:
+            # In this case, we must stop CentralReport immediatly!
+            if not self.SIGTERM_SENT:
+                self.SIGTERM_SENT = True  # Prevents if SIGTERM is received twice.
 
-            log.log_info('SIGTERM signal received. Shutting Down...')
-            log.log_info('Shutting down sub-processes...')
-            os.killpg(0, signal.SIGTERM)
+                log.log_info('SIGTERM signal received (%s). Shutting Down...' % signum)
+                log.log_info('Shutting down sub-processes...')
+                os.killpg(0, signal.SIGTERM)
 
+                self.stop()
+
+        elif signum == signal.SIGINT:
+            # Keyboard interruption (CTRL + C)
+            log.log_info('SIGINT signal received (%s). Stopping CentralReport...' % signum)
             self.stop()
 
     def run(self):
@@ -54,12 +61,14 @@ class CentralReport(Daemon):
 
         is_error = False  # If True, there are one or more errors when CentralReport is trying to start
 
+        log.log_info('------------------------------------------------')
         log.log_info('CentralReport is starting...')
         log.log_info('Current user: ' + getpass.getuser())
 
         # Registring SIGTERM signal event
         os.setsid()
-        signal.signal(signal.SIGTERM, self.sigterm_handler)
+        signal.signal(signal.SIGTERM, self.signal_handler)
+        signal.signal(signal.SIGINT, self.signal_handler)
 
         CentralReport.starting_date = datetime.datetime.now()  # Starting date
         CentralReport.configuration = Config()  # Getting config object
@@ -77,10 +86,10 @@ class CentralReport(Daemon):
         if not is_error and text.convert_text_to_bool(Config.get_config_value('Webserver', 'enable')):
             from web.server import WebServer
 
-            log.log_info('Enabling the webserver')
+            log.log_info('Enabling the webserver...')
             CentralReport.webserver_thread = WebServer()
         else:
-            log.log_info('Webserver is disabled by configuration file')
+            log.log_info('Webserver is disabled by configuration file!')
 
         if not is_error:
             log.log_info('CentralReport started!')
@@ -99,9 +108,8 @@ class CentralReport(Daemon):
                     time.sleep(1)
 
                 except KeyboardInterrupt:
-                    # Stopping CR
-                    log.log_critical('KeyboardInterrupt exception. Stopping CentralReport...')
                     CentralReport.is_running = False
+                    log.log_warning('KeyboardInterrupt exception!')
                     self.stop()
         else:
             log.log_error('Error launching CentralReport!')
@@ -124,13 +132,9 @@ class CentralReport(Daemon):
 
         log.log_info('Stopping daemon...')
 
-        # In test mode, we only return 0 (exit can be personalized by others scripts)
-        # But in production, we kill immediately the process.
-
-        if not Config.CR_CONFIG_ENABLE_DEBUG_MODE:
-            os.system('kill -9 %d' % os.getpid())
-
-        return 0
+        # Killing the current processus...
+        # (This command send the "SIGKILL" signal)
+        os.system('kill -9 %d' % os.getpid())
 
     def status(self):
         """
